chore: improve thread ux (#347)
* chore: improve thread ux

* feat: improve drag handling
