+++ apps/web/package.json
@@ -11,6 +11,9 @@
   "dependencies": {
     "@c15t/nextjs": "^2.0.4",
     "@databuddy/sdk": "^2.4.11",
+    "@dnd-kit/core": "^6.3.1",
+    "@dnd-kit/sortable": "^10.0.0",
+    "@dnd-kit/utilities": "^3.2.2",
     "@dualmark/converters": "0.5.0",
     "@dualmark/nextjs": "0.5.1",
     "@hugeicons/core-free-icons": "^4.1.1",
